refactor(exception): 폴더 변경 (common 폴더로 이동)
공통된 layer에서 사용할 수 있다고 생각해서 폴더를 이동함.
